Add answers to mongoose-crud-diagnostic
Comfort: 3
Clarity: 3

I used Google and my class notes to help with this diagnostic.  I didn't have
enough time to finish.  But I feel somewhat comfortable with this. I definitely
had to follow along with yesterdays work.  I'm not very sure on how to test the work.
I also don't see why we needed to create a new database, I'm not even sure mine
worked by typing in the command to do so.  I need more help with trying to understand
the testing aspect of this for sure.